update discord code for better readability
